DDPCM improvements (#29)
* update encoder to use inverse fixed point scales.

* fix build on linux

* update ddpcm code to work with new format.

* tidy up

* tidy up

* update include

* add newly encoded audio assets.

* Use ahead of time unpacking until realtime decoding is fixed.

+++ include/ddpcm_lowlevel.h
@@ -42,7 +42,7 @@ void *getDDPCMMixRoutine16(void);
 void decodeFrame_asm(__ASMPARM("a0", uint8_t *src),
                      __ASMPARM("a6", int16_t *dst),
                      __ASMPARM("a5", int16_t *q_table),
-                     __ASMPARM("d7", uint8_t scale));
+                     __ASMPARM("d7", int32_t scale));
 
 void ddpcmAHIPlayerFunc(__ASMPARM("d0", uint32_t numSamples),
                         __ASMPARM("a0", uint16_t *left),
